requirements.txt
streamlit
torch
transformers
ultralytics
google-genai
python-dotenv